:zap: Improve database access logic
This changes the database access logic so that we only query it when needed.

The whole config system was changed and, instead of loading the user
config on a middleware, now we only do it when `ctx.config.getAll`,
`ctx.config.get` or `ctx.config.set` are called.

There's also a 10 minute cache for the config data.